Updates as per released FDO spec (#36)
There were some updates to the 1.0 spec as compared
to RD spec. Following changes were updated to comply with
new spec definition.

- Updating SHA256 and SHA384 constants.
- Updating the names of Nonces.

+++ crypto/openssl/openssl_cryptoSupport.c
@@ -244,7 +244,7 @@ int32_t crypto_hal_hash(uint8_t _hash_type, const uint8_t *buffer,
 			 size_t buffer_length, uint8_t *output,
 			 size_t output_length)
 {
-	uint8_t hash_type = FDO_CRYPTO_HASH_TYPE_USED;
+	int hash_type = FDO_CRYPTO_HASH_TYPE_USED;
 
 	(void)_hash_type; /* Unused parameter */
